Green NFT
It takes 1 ACRE of forest 1 year to sequester the CO2 impact of just 63 ETH transactions. An Ethereum transaction contributes about 79lbs of CO2 to the atmosphere, and an acre of trees will remove about 2.5 tons of carbon annually. The annualized total Ethereum Carbon Footprint is about 18,620,00 Millions Tons of CO2, so we need 7,448,000 ACRES of forest to sequester the carbon footprint of the ETH network.